[svn-r6515] Purpose: Bug fix
Description:
The $RUNPARALLEL is set even when parallel mode is not used.
Then it is used to launch some parallel natured test (e.g. example/ph5example)
which is not compiled to be so.  That created some unnecessary problem
in some situation (e.g., Copper cannot "poe a.out" from an AFS space
because of the missing AFS token system bug.

Solution:
Setup RUNPARALLEL only when parallel mode is enabled.

Platforms tested:
Only tested in Copper since the change applies to IBM machines only.
(note that v1.4 will fail by default due to a problem when the default
production mode is used.  But that is an error not related to this config file)

Misc. update: