[svn-r6716] Purpose: Typo fix.
Description:
Found a typo in the enable-asyncio processing.  Did not get exposed
unless you ran it in an asyncio working system that uses a zlib that is
installed in a non-system area.

Solution:
Fixed the typo in CPPFLAGS.

Platforms tested:
Tested at LLNL Blue only--that is the only system that asyncio works
currently.  Pretty sure the fix won't break other systems.

Misc. update: