[svn-r6782] Purpose: Bug fix
Description:
"$@" was used to represent all argument provided.  This has the
unfortunate effect as one argument of "" when actually no argument
was given.  Some machines throw the "" away but some like Tflops
and Alpha Q, retains them as a single argument of "".  This caused
the no filename given test to fail.

Solution:
Replaced "$@" with a plain $@.

Platforms tested:
h5committested

Misc. update: