[svn-r27724] Checkin to correct oversight in failing to add src/H5Omdci_msg.c to the repository.
Also added test/genall5.c test/genall5.h to the repository, along with 
required changes to test/Makefile.am and test/Makefile.in to get them
to build with the tests.  Note that the code in genall5.c was written 
by Quincey to assist with cache image smoke checks.  These two files
are very much in preliminary form.

Tested serial on Jam.