Checkin of merge of file shutdown modifications into page buffering.
Note that while page buffering passes its tests, it still needs a
shave and a haircut before merge into trunk.

Also note that this checkin contains one know assertion failure in
test/mf.  Specifically the paged allocation H5MF_alloc/H5MF_xfree
test fails on file shutdown when the EOA is not at a page boundary.
This may or may not be related to the file shutdown changes, but I
am passing the issue to Vailin as she should be better prepared to
chase the issue than I.

Finally, note that the code is not in a fit state for final merge
with Develop.  I will tidy up when I review the page buffering code.

Tested serial debug and production on Mercury
       serial debug on Jelly