Added code to resolve remaining corner cases in integration of cache image with file shutdown fix as modified for paged allocation.
Added test to verify correctness of fix for H5MF_get_free_sections()
case.

Used the same fix for H5MF_try_close(), but did not write test code
for it.  It may be appropriate to do this at some point in the
future.

Tested on Jelly serial debug and production, and parallel debug.